+++ src/main/java/com/example/auction/common/exception/ErrorCode.java
@@ -1,7 +1,5 @@
 package com.example.auction.common.exception;
 
-import org.springframework.http.HttpStatus;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -20,13 +18,9 @@ public enum ErrorCode implements BaseCode {
 	//DB
 	DB_LOCK_CONFLICT(HttpStatus.CONFLICT, "L002", "데이터베이스 락 상태로 요청을 처리할 수 없습니다."),
 
-	//JWT
-	INVALID_JWT_SIGNATURE(HttpStatus.UNAUTHORIZED, "A001", "유효하지 않은 JWT 서명입니다."),
-	EXPIRED_JWT_TOKEN(HttpStatus.UNAUTHORIZED, "A002", "만료된 JWT 토큰입니다."),
-	UNSUPPORTED_JWT_TOKEN(HttpStatus.BAD_REQUEST, "A003", "지원되지 않는 JWT 토큰입니다."),
-	INVALID_JWT_TOKEN(HttpStatus.BAD_REQUEST, "A004", "잘못된 JWT 토큰입니다."),
-	UNKNOWN_JWT_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "A005", "알 수 없는 JWT 관련 오류가 발생했습니다."),
-	NOT_FOUND_JWT_TOKEN(HttpStatus.NOT_FOUND, "A006", "존재하지않는 JWT 토큰입니다."),
+	//AUTH
+	INVALID_SIGNATURE(HttpStatus.UNAUTHORIZED, "A001", "인증되지 않은 사용자입니다."),
+	UNAUTHOIZED(HttpStatus.UNAUTHORIZED, "A002", "접근 권한이 없습니다."),
 
 	//ImageAspect
 	INVALID_IMAGE_SIZE(HttpStatus.BAD_REQUEST, "I001", "이미지 파일의 크기는 최대 2MB까지 업로드 가능합니다."),
