+++ src/main/java/com/example/auction/domain/product/controller/ProductController.java
@@ -2,6 +2,10 @@ package com.example.auction.domain.product.controller;
 
 import java.util.List;
 
+import jakarta.validation.Valid;
+
+import lombok.RequiredArgsConstructor;
+
 import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -19,7 +23,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.example.auction.common.annotation.ImageValid;
 import com.example.auction.common.response.CommonResponse;
-import com.example.auction.domain.auth.security.CustomUserDetails;
 import com.example.auction.domain.product.dto.request.ProductRequestDto;
 import com.example.auction.domain.product.dto.request.ProductUpdateRequestDto;
 import com.example.auction.domain.product.dto.response.PageResponseDto;
@@ -27,9 +30,7 @@ import com.example.auction.domain.product.dto.response.ProductResponseDto;
 import com.example.auction.domain.product.dto.response.ProductSaveResponseDto;
 import com.example.auction.domain.product.dto.response.ProductWithdrawResponseDto;
 import com.example.auction.domain.product.service.ProductService;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
+import com.example.auction.domain.user.auth.security.CustomUserDetails;
 
 @RestController
 @RequestMapping("/api/products")
