+++ src/main/java/com/example/auction/domain/topKeyword/controller/TopKeywordController.java
@@ -2,17 +2,17 @@ package com.example.auction.domain.topKeyword.controller;
 
 import java.util.List;
 
+import lombok.RequiredArgsConstructor;
+
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.auction.common.response.CommonResponse;
-import com.example.auction.domain.auth.security.CustomUserDetails;
 import com.example.auction.domain.topKeyword.dto.response.TopKeywordResponseDto;
 import com.example.auction.domain.topKeyword.service.TopKeywordService;
-
-import lombok.RequiredArgsConstructor;
+import com.example.auction.domain.user.auth.security.CustomUserDetails;
 
 @RestController
 @RequestMapping("/api/popular-keywords")
