+++ src/main/resources/application.properties
@@ -17,6 +17,8 @@ spring.data.redis.host=${REDIS_HOST}
 spring.data.redis.port=${REDIS_PORT}
 #image
 app.default-image-id=1
+#jackson
+spring.jackson.serialization.WRITE_DATES_AS_TIMESTAMPS=false
 #social login
 # =google ===
 spring.security.oauth2.client.registration.google.client-id=${GOOGLE_ID}
